pyproject.toml
[tool.poetry]
name = "groovis"
version = "0.1.0"
description = ""
authors = ["Kwang Myung Yu <sguys99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^10.1.0"
torch = "^1.12.0"
numpy = "^1.26.2"
einops = "^0.7.0"
albumentations = "^1.4.0"
matplotlib = "^3.8.3"
torchvision = "^0.13.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.10.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
